+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\index.js":"1","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\App.js":"2","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\reportWebVitals.js":"3","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\allPlants\\AllPlants.js":"4","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\components\\header\\Header.js":"5","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\aboutUs\\AboutUs.js":"6","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\home\\Home.js":"7","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\context\\LanguageContext.js":"8"},{"size":517,"mtime":1617902269973,"results":"9","hashOfConfig":"10"},{"size":1267,"mtime":1617915244236,"results":"11","hashOfConfig":"10"},{"size":375,"mtime":1617902269975,"results":"12","hashOfConfig":"10"},{"size":555,"mtime":1617914753430,"results":"13","hashOfConfig":"10"},{"size":1994,"mtime":1617916571672,"results":"14","hashOfConfig":"10"},{"size":637,"mtime":1617914753434,"results":"15","hashOfConfig":"10"},{"size":439,"mtime":1617914753439,"results":"16","hashOfConfig":"10"},{"size":89,"mtime":1617914218438,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"fn004x",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\index.js",[],["35","36"],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\App.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\reportWebVitals.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\allPlants\\AllPlants.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\components\\header\\Header.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\aboutUs\\AboutUs.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\home\\Home.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\context\\LanguageContext.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\index.js":"1","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\App.js":"2","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\reportWebVitals.js":"3","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\allPlants\\AllPlants.js":"4","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\components\\header\\Header.js":"5","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\aboutUs\\AboutUs.js":"6","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\home\\Home.js":"7","C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\context\\LanguageContext.js":"8"},{"size":667,"mtime":1617917360505,"results":"9","hashOfConfig":"10"},{"size":840,"mtime":1617917481450,"results":"11","hashOfConfig":"10"},{"size":375,"mtime":1617902269975,"results":"12","hashOfConfig":"10"},{"size":555,"mtime":1617917500274,"results":"13","hashOfConfig":"10"},{"size":1994,"mtime":1617917038293,"results":"14","hashOfConfig":"10"},{"size":637,"mtime":1617917048954,"results":"15","hashOfConfig":"10"},{"size":439,"mtime":1617917493216,"results":"16","hashOfConfig":"10"},{"size":602,"mtime":1617917525063,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"fn004x",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\index.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\App.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\reportWebVitals.js",[],["35","36"],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\allPlants\\AllPlants.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\components\\header\\Header.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\aboutUs\\AboutUs.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\pages\\home\\Home.js",[],"C:\\Users\\chris\\WebstormProjects\\frontend-react-language-context\\src\\context\\LanguageContext.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
